+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CharacterFlag Multiplier Selector
 - Presets
 - Separators for Capacity UI
+- FlagMultipliers For Tags
+- Health and Level Scaling
+
+### Fixed
+- Mid air grab for capping your capacity
+- Capacity not being set to 1 when grabbing a big object
+- UncapCapacity + MassCapacity not capping correctly
 
 ## [1.7.0]